chore: put bash autocomplete scripts in the own files
